Await prior command completion in command sequences
This is a refactor of the way that command sequences
(KeymapResult::MatchedSequence) are handled. This fixes some scenarios
where commands use on-next-key callbacks.

Consider these example mappings:

    D = ["find_till_char", "@<ret>", "delete_selection"]
    F = ["find_till_char", "collapse_selection"]

In the first, the macro `<ret>` key should apply to the
`find_till_char` and `delete_selection` should act on the selection
created by the first two commands. Prior to this change though, the
`delete_selection` would operate immediately and then the other two
commands would act.

In the second, `collapse_selection` should await the selection created
by `find_till_char`'s on-next-key callback. Prior to this change,
`collapse_selection` acted first and then `find_till_char`.

In this change, we add a fake event to the compositor which we emit
when we finish executing a command in a sequence.

When handling a sequence, we:

1. naively execute the first command.
2. store the remaining commands on the EditorView.
3. setup a callback to emit the new event once done handling the
   current command.

To handle the second case, we also emit the new fake event when we
finish executing a KeyEvent but have outstanding pending commands.